- Remove module type script
