+++ index.html
@@ -175,6 +175,6 @@
       </div>
     </section>
     <div id="modal-container"></div>
-    <script type="module" src="./js/main.js"></script>
+    <script src="./js/main.js"></script>
   </body>
 </html>
