--- js/cardData.js
@@ -1,60 +1,0 @@
-const cardData = [
-  {
-    title: 'Tonic',
-    isReversed: false,
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.",
-    longDescription:
-      "<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent</p>",
-    workImg: './img/snapshoot_portfolio.png',
-    workImgDesk: './img/desktop-snapshot.svg',
-    details: ['Canopy', 'Back End Dev', '2015'],
-    tags: ['HTML', 'CSS', 'JavaScript'],
-    demoURL: 'https://www.google.com',
-    githubURL: 'https://www.github.com',
-  },
-  {
-    title: 'Multi-Post Stories',
-    isReversed: true,
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.",
-    longDescription:
-      "<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent</p>",
-    workImg: './img/snapshoot-profesional.png',
-    workImgDesk: './img/desktop-snapshot2.svg',
-    details: ['Canopy', 'Back End Dev', '2015'],
-    tags: ['HTML', 'Ruby', 'CSS', 'JavaScript'],
-    demoURL: 'https://www.google.com',
-    githubURL: 'https://www.github.com',
-  },
-  {
-    title: 'Facebook 360',
-    isReversed: false,
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.",
-    longDescription:
-      "<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent</p>",
-    workImg: './img/snapshoot-portfolio.png',
-    workImgDesk: './img/desktop-snapshot3.svg',
-    details: ['Canopy', 'Back End Dev', '2015'],
-    tags: ['HTML', 'Ruby', 'CSS', 'JavaScript'],
-    demoURL: 'https://www.google.com',
-    githubURL: 'https://www.github.com',
-  },
-  {
-    title: 'Uber Navigation',
-    isReversed: true,
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.",
-    longDescription:
-      "<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent</p>",
-    workImg: './img/snapshoot-portfolio2.png',
-    workImgDesk: './img/desktop-snapshot4.svg',
-    details: ['Uber', 'Back End Dev', '2015'],
-    tags: ['HTML', 'Ruby', 'CSS', 'JavaScript'],
-    demoURL: 'https://www.google.com',
-    githubURL: 'https://www.github.com',
-  },
-];
-
-export default cardData;
